feat: maturity-docs-recording-002 - Explain equivalent execution surfaces and worker presets
