+++ pkg/orchestrators/javascript/runtime/runtime_test.go
@@ -103,91 +103,6 @@ func TestRun_SimpleFinalWorkflow_ProjectsStructuredPrimaryResult(t *testing.T) {
 	assertProjectedFields(t, projected, want)
 }
 
-func TestRun_DocumentedJavaScriptWorkflowExamples(t *testing.T) {
-	t.Run("simple final", func(t *testing.T) {
-		outcome := runDocumentedExample(t, "simple-final.workflow.js", map[string]any{
-			"greeting": "hello", "subject": "factory authors",
-		}, nil)
-		projected := projectPrimaryJSON(t, "session-documented-simple-final", outcome.Value)
-		if projected["greeting"] != "hello" || projected["subject"] != "factory authors" {
-			t.Fatalf("documented simple result = %#v", projected)
-		}
-		if len(outcome.Records) != 1 || outcome.Records[0].Kind != workflowruntime.RecordKindPhase {
-			t.Fatalf("documented simple records = %#v", outcome.Records)
-		}
-	})
-
-	t.Run("ordered fanout with synthesis", func(t *testing.T) {
-		outcome := runDocumentedExample(t, "ordered-fanout.workflow.js", map[string]any{
-			"topics": []string{"alpha", "beta", "gamma"},
-		}, nil)
-		projected := projectPrimaryJSON(t, "session-documented-ordered-fanout", outcome.Value)
-		reviews, ok := projected["reviews"].([]any)
-		if !ok || len(reviews) != 3 {
-			t.Fatalf("documented reviews = %#v", projected["reviews"])
-		}
-		for index, topic := range []string{"alpha", "beta", "gamma"} {
-			review, ok := reviews[index].(map[string]any)
-			if !ok || review["label"] != "review-"+topic {
-				t.Fatalf("documented reviews[%d] = %#v", index, reviews[index])
-			}
-		}
-		if synthesis, ok := projected["synthesis"].(map[string]any); !ok || synthesis["label"] != "synthesize" {
-			t.Fatalf("documented synthesis = %#v", projected["synthesis"])
-		}
-		if childTerminalRecordCount(outcome.Records) != 4 {
-			t.Fatalf("completed child dispatches = %d, want 4", childTerminalRecordCount(outcome.Records))
-		}
-	})
-
-	t.Run("checkpoint and resume", func(t *testing.T) {
-		args := map[string]any{"topics": []string{"alpha", "beta"}}
-		fresh := runDocumentedExample(t, "checkpoint-resume.workflow.js", args, nil)
-		var checkpoint *workflowruntime.CheckpointRecord
-		for _, record := range fresh.Records {
-			if record.Checkpoint != nil {
-				checkpoint = record.Checkpoint
-			}
-		}
-		if checkpoint == nil || checkpoint.State["completedTopics"] == nil {
-			t.Fatalf("documented checkpoint records = %#v", fresh.Records)
-		}
-		resumed := runDocumentedExample(t, "checkpoint-resume.workflow.js", args, &workflowruntime.ResumeContext{
-			CheckpointState: checkpoint.State,
-		})
-		projected := projectPrimaryJSON(t, "session-documented-checkpoint-resume", resumed.Value)
-		if projected["path"] != "resumed" {
-			t.Fatalf("documented resumed result = %#v", projected)
-		}
-	})
-}
-
-func runDocumentedExample(t *testing.T, name string, args map[string]any, resume *workflowruntime.ResumeContext) workflowruntime.Outcome {
-	t.Helper()
-	path := filepath.Join("..", "..", "..", "..", "docs", "examples", "javascript-workflows", name)
-	source, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatalf("read documented workflow %s: %v", name, err)
-	}
-	return runSuccessful(t, workflowruntime.Request{
-		Source: string(source), SourceRef: path,
-		SessionID: "session-documented-" + strings.TrimSuffix(name, ".workflow.js"),
-		Args:      marshalArgs(t, args),
-		Metadata:  map[string]string{"name": strings.TrimSuffix(name, ".workflow.js")},
-		Policy:    workflowpolicy.DefaultEffectivePolicy(), Resume: resume,
-	})
-}
-
-func childTerminalRecordCount(records []workflowruntime.RuntimeRecord) int {
-	count := 0
-	for _, record := range records {
-		if record.ChildDispatch != nil && record.ChildDispatch.Status == workflowruntime.ChildDispatchStatusCompleted {
-			count++
-		}
-	}
-	return count
-}
-
 func TestRun_WorkflowFinal_ProjectsStructuredPrimaryResult(t *testing.T) {
 	source := readFixture(t, "workflow-final.workflow.js")
 	args := marshalArgs(t, map[string]any{
